+++ src/creatures/appearance/outfit/outfit.hpp
@@ -73,7 +73,7 @@ public:
 	[[nodiscard]] const std::vector<std::shared_ptr<Outfit>> &getOutfits(PlayerSex_t sex) const;
 
 	std::shared_ptr<Outfit> getOutfitByName(PlayerSex_t sex, const std::string_view &name) const;
-	uint32_t getOutfitId(PlayerSex_t sex, uint32_t lookType) const;
+	uint32_t getOutfitId(PlayerSex_t sex, uint16_t lookType) const;
 
 	bool addAttributes(uint32_t playerId, uint32_t outfitId, uint16_t sex, uint16_t addons);
 	bool removeAttributes(uint32_t playerId, uint32_t outfitId, uint16_t sex);
